+++ test/unabto/util/unabto_buffer_test.c
@@ -6,7 +6,7 @@ bool unabto_buffer_test(void)
 {
     char * raw_test_string = "asjdhc#21?(!?(92814skzjbcasa";
 
-    int data_length = 7 + 2 + strlen(raw_test_string);
+    size_t data_length = 7 + 2 + strlen(raw_test_string);
     uint8_t* data = (uint8_t*)malloc(data_length);
     unabto_buffer buf, raw_string_buf;
     buffer_write_t w_buf;
@@ -15,7 +15,7 @@ bool unabto_buffer_test(void)
     uint32_t t32;
     uint16_t t16;
     uint8_t  t8;
-    int raw_string_data_length = strlen(raw_test_string) + 1;
+    size_t raw_string_data_length = strlen(raw_test_string) + 1;
     uint8_t* raw_string_data = (uint8_t*)malloc(raw_string_data_length);
 
     buffer_init(&raw_string_buf, (uint8_t*)raw_test_string, strlen(raw_test_string));
